Added cbot_monitoring package
